requirements.txt
pyTelegramBotAPI
google-generativeai